Testing adding a Network Observer Component to P_Player
Testing a change to the P_Player prefab (part of VR2G Core). I found a mention, when trawling through Fishnet tutorials, that the player controlled objects need to have a Network Observer component to interact and see dynamic networked objects (any object with behaviours). I figured this might be related to the issue we were having with the button earlier.